String

+++ printExclusive_string.c
@@ -8,7 +8,7 @@
  * @val: argument.
  * Return: string length
  */
-int printHEX_aux(unsigned long int num);
+int printHEX_aux(unsigned int num);
 
 int printExclusive_string(va_list val)
 {
